Alternating Positives and Negatives
Given an array named array of size n, that contains both positive and negative numbers. Rearrange the array elements so that positive and negative numbers appear alternatively in the output. The order in which the positive elements appear should be maintained. Similarly, the order in which the negative elements appear should also be maintained.

Number of positive and negative integers may not be equal and extra positives or negatives have to appear in the end of the array.